fix: remove auth from `GET /caches` route
